Promote Simpler sample to top-level field on create/update-device
The pseudo-param-in-params approach was bloating the params description
with device-class-specific text and inconsistent with the codebase's
established pattern: device-class-specific writes (macroVariation,
chokeGroup, mappedPitch, abCompare) are top-level fields with
warnIfSet, not params entries. params is for DeviceParameter resolution.

Restructured to match:
- New top-level sample field on update-device and create-device
- Shared setSimplerSample helper handles class_display_name check,
  multi_sample_mode skip, and replace_sample call for both flows
- Reverted params description and pseudo-param dispatcher
- Extracted updateDeviceProperties/updateNonDeviceProperties to a new
  helper module (update-device.ts was approaching the 325-line limit)
