+++ src/skills/standard.ts
@@ -200,7 +200,7 @@ Slash-separated segments: \`t\`=track, \`rt\`=return, \`mt\`=master, \`d\`=devic
 
 Chains are auto-created when referenced (e.g., \`c0\` on an empty rack creates a chain). Up to 16 chains.
 
-**Simpler sample:** Update via \`params\` with \`sample=<absolute file path>\` to load a sample into a Simpler device.
+**Simpler sample:** Pass \`sample=<absolute file path>\` to ppal-create-device or ppal-update-device to load a sample. Skipped with a warning on non-Simpler devices and on Simpler in multi-sample mode.
 
 ### Moving Clips
