+++ src/tools/device/create/create-device.def.ts
@@ -28,6 +28,12 @@ export const toolDefCreateDevice = defineTool("ppal-create-device", {
       .string()
       .optional()
       .describe("name for all, or comma-separated for each"),
+    sample: z
+      .string()
+      .optional()
+      .describe(
+        "Simpler only: absolute file path to load as the sample (after creation)",
+      ),
   },
 
   smallModelModeConfig: {
@@ -35,6 +41,7 @@ export const toolDefCreateDevice = defineTool("ppal-create-device", {
     descriptionOverrides: {
       path: "insertion path, required with deviceName (e.g., 't0', 't0/d1', 't0/d0/c0')",
       name: "display name",
+      sample: "Simpler only: file path to load",
     },
   },
 });
