+++ src/tools/device/create/create-device.ts
@@ -7,6 +7,7 @@ import * as console from "#src/shared/v8-max-console.ts";
 import { ALL_VALID_DEVICES, VALID_DEVICES } from "#src/tools/constants.ts";
 import { select } from "#src/tools/control/select.ts";
 import { resolveInsertionPath } from "#src/tools/shared/device/helpers/path/device-path-helpers.ts";
+import { setSimplerSample } from "#src/tools/shared/device/simpler-sample.ts";
 import {
   parseCommaSeparatedIds,
   unwrapSingleResult,
@@ -21,6 +22,7 @@ interface CreateDeviceArgs {
   deviceName?: string;
   path?: string;
   name?: string;
+  sample?: string;
   focus?: boolean;
 }
 
@@ -54,12 +56,13 @@ function validateDeviceName(deviceName: string): void {
  * @param args.deviceName - Device name, omit to list available devices
  * @param args.path - Device path(s), comma-separated for multiple (required when deviceName provided)
  * @param args.name - Name for all, or comma-separated for each
+ * @param args.sample - Simpler only: file path to load after creation
  * @param args.focus - Select the device and show device detail view
  * @param _context - Internal context object (unused)
  * @returns Device list, or object(s) with deviceId and deviceIndex
  */
 export function createDevice(
-  { deviceName, path, name, focus }: CreateDeviceArgs = {},
+  { deviceName, path, name, sample, focus }: CreateDeviceArgs = {},
   _context: Partial<ToolContext> = {},
 ): typeof VALID_DEVICES | CreateDeviceResult | CreateDeviceResult[] {
   // List mode: return valid devices when deviceName is omitted
@@ -81,7 +84,13 @@ export function createDevice(
 
   warnExtraNames(parsedNames, paths.length, "createDevice");
 
-  const results = createDevicesAtPaths(deviceName, paths, name, parsedNames);
+  const results = createDevicesAtPaths(
+    deviceName,
+    paths,
+    name,
+    parsedNames,
+    sample,
+  );
 
   if (focus && results.length > 0) {
     const lastResult = results.at(-1) as CreateDeviceResult;
@@ -98,6 +107,7 @@ export function createDevice(
  * @param paths - Array of device paths
  * @param baseName - Base display name
  * @param parsedNames - Comma-separated display names, or null
+ * @param sample - Simpler only: file path to load after creation
  * @returns Array of results for successfully created devices
  */
 function createDevicesAtPaths(
@@ -105,6 +115,7 @@ function createDevicesAtPaths(
   paths: string[],
   baseName: string | undefined,
   parsedNames: string[] | null,
+  sample: string | undefined,
 ): CreateDeviceResult[] {
   const results: CreateDeviceResult[] = [];
 
@@ -113,14 +124,15 @@ function createDevicesAtPaths(
 
     try {
       const result = createDeviceAtPath(deviceName, p);
+      const device = LiveAPI.from(`id ${result.id}`);
       const displayName = getNameForIndex(baseName, i, parsedNames);
 
-      if (displayName != null) {
-        const device = LiveAPI.from(`id ${result.id}`);
+      if (displayName != null && device.exists()) {
+        device.set("name", displayName);
+      }
 
-        if (device.exists()) {
-          device.set("name", displayName);
-        }
+      if (sample != null && device.exists()) {
+        setSimplerSample(device, sample, "createDevice");
       }
 
       results.push(result);
